Add ownership checks for review update and delete (#44)

+++ src/main/java/api/indy/kebab/service/ReviewService.java
@@ -1,6 +1,8 @@
 package api.indy.kebab.service;
 
+import api.indy.kebab.auth.Permission;
 import api.indy.kebab.exceptions.EntityNotFoundException;
+import api.indy.kebab.exceptions.NotOwnerOfEntityException;
 import api.indy.kebab.model.Review;
 import api.indy.kebab.model.User;
 import api.indy.kebab.repository.ReviewRepository;
@@ -109,9 +111,14 @@ public class ReviewService {
      * @throws IOException if an I/O error occurs during icon upload.
      * @throws EntityNotFoundException if the review with the specified ID does not exist.
      */
-    public Review updateReview(long id, String title, String description, MultipartFile image, Double rating, Boolean anonymous) throws IOException {
+    public Review updateReview(
+            HttpSession session, long id, String title, String description, MultipartFile image, Double rating, Boolean anonymous
+    ) throws NotOwnerOfEntityException, IOException {
         Review review = this._reviewRepository.findByReviewId(id);
         if(review == null) throw new EntityNotFoundException(Review.class, id);
+
+        if(!Util.verifyOwnership(review.getUserId(), Permission.NONE, this._authService.getActiveUser(session)))
+            throw new NotOwnerOfEntityException(Review.class);
 
         if(title != null && !title.isEmpty()) review.setTitle(title);
         if(description != null && !description.isEmpty()) review.setDescription(description);
@@ -136,9 +143,12 @@ public class ReviewService {
      *
      * @param id the unique identifier of the review to delete.
      */
-    public void deleteReview(long id) {
+    public void deleteReview(HttpSession session, long id) throws NotOwnerOfEntityException {
         Review review = this._reviewRepository.findByReviewId(id);
         if(review == null) throw new EntityNotFoundException(Review.class, id);
+
+        if(!Util.verifyOwnership(review.getUserId(), Permission.REVIEWS_DELETE, this._authService.getActiveUser(session)))
+            throw new NotOwnerOfEntityException(Review.class);
 
         Util.deleteFile(review.getImageUrl());
         this._reviewRepository.delete(review);
@@ -149,9 +159,12 @@ public class ReviewService {
      *
      * @param id the unique identifier of the review whose image is to be deleted.
      */
-    public void deleteReviewImage(long id) {
+    public void deleteReviewImage(HttpSession session, long id) throws NotOwnerOfEntityException {
         Review review = this._reviewRepository.findByReviewId(id);
         if(review == null) throw new EntityNotFoundException(Review.class, id);
+
+        if(!Util.verifyOwnership(review.getUserId(), Permission.REVIEWS_UPDATE, this._authService.getActiveUser(session)))
+            throw new NotOwnerOfEntityException(Review.class);
 
         Util.deleteFile(review.getImageUrl());
         review.setImageUrl(null);
